// src/movies/movies.model.d.ts
export interface movieDTO {
    id: number;
    title: string;
    poster: string;
}

export interface movieCreationDTO{
    title: string;
    inTheaters: boolean;
    trailer: string;
    releaseDate?: Date;
    poster?: File;
    porterURL?: string;
}

export interface landingPageDTO{
    inTheaters?: movieDTO[];
    upComingReleases?: movieDTO[];
}